Ran Static code tool Pylint: 10/10
A job description wanted static tool analysis experience, so I decided to run Pylint on this program and get it all sexied up. My biggest offender was bad whitespace lol. Thanks Python. Reordered a couple imports, apparently using upper case for variables outside of a scope is a thing, and I included a docstring.

Score is 10/10 on Pylint :)